+++ R/liana_plot.R
@@ -48,7 +48,8 @@ liana_dotplot <- function(liana_res,
                           size.label = "Interaction\nSpecificity",
                           colour.label = "Expression\nMagnitude",
                           show_complex = TRUE,
-                          size_range = c(2, 10)){
+                          size_range = c(2, 10)
+                          ){
 
     if(show_complex){
         entities <- c("ligand.complex", "receptor.complex")
@@ -83,7 +84,7 @@ liana_dotplot <- function(liana_res,
     # colour blind palette from http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/
     cbPalette <- c("#E69F00", "#56B4E9",
                    "#009E73", "#F0E442", "#0072B2",
-                   "#D55E00", "#CC79A7")
+                   "#D55E00", "#CC79A7", "#DF69A7")
 
     # plot
     suppressWarnings(
@@ -100,7 +101,14 @@ liana_dotplot <- function(liana_res,
             facet_grid(. ~ source,
                        space = "free",
                        scales ="free",
-                       switch="y") +
+                       switch="y")  +
+            # scale_x_discrete(position = "right") +
+            labs(y = y.label,
+                 colour = colour.label,
+                 size = size.label,
+                 x = "Target",
+                 title= "Source"
+            ) +
             theme_bw(base_size = 20) +
             theme(
                 legend.text = element_text(size = 16),
@@ -110,19 +118,15 @@ liana_dotplot <- function(liana_res,
                                                )],
                                            face = "bold",
                                            size = 23),
+                axis.title.x = element_text(colour = "gray6"),
                 axis.text.y = element_text(size = 18,
                                            vjust = 0.5),
                 legend.title = element_text(size = 18),
                 panel.spacing = unit(0.1, "lines"),
                 strip.background = element_rect(fill = NA),
+                plot.title = element_text(vjust = 0, hjust=0.5, colour = "gray6"),
                 strip.text = element_text(size = 24, colour = "gray6") #,
                 # strip.text.y.left = element_text(angle = 0)
-            ) +
-            # scale_x_discrete(position = "right") +
-            labs(y = y.label,
-                 colour = colour.label,
-                 size = size.label,
-                 x = NULL
             )
     )
 }